refactor(services): slim cell/content/mod.rs by extracting siblings
Split the 734-line mod.rs into engine_loader (DB chain loading), event_dispatch
(fire_* functions), and mission_context (per-entity context helper). mod.rs is
now a thin glue file (190 lines) that declares submodules and re-exports the
public surface.

Public API unchanged. No behavior changes.
