+++ crates/services/src/cell/content/event_dispatch.rs
@@ -1,0 +1,426 @@
+//! Event firing — gameplay sites call into these functions to push events
+//! through the content engine and execute any matched actions.
+//!
+//! Each `fire_*` function:
+//!   1. Builds an [`ExecutionContext`] populated with event-specific params
+//!      and the source entity's mission/archetype state.
+//!   2. Constructs a [`TriggerEvent`] and resolves it against the engine.
+//!   3. Dispatches the resolved actions via [`super::executor::execute_actions`].
+
+use tokio::sync::mpsc;
+
+use cimmeria_content_engine::chain::ChainEngine;
+use cimmeria_content_engine::context::ExecutionContext;
+use cimmeria_content_engine::triggers::{TriggerEvent, TriggerType};
+
+use crate::cell::messages::CellToBaseMsg;
+use crate::cell::space_manager::SpaceManager;
+
+use super::executor;
+use super::mission_context::populate_mission_context;
+
+/// Fire the `PlayerLoaded` event for a player entering a world.
+pub async fn fire_player_loaded(
+    entity_id: u32,
+    player_id: i32,
+    world_name: &str,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+
+    ctx.set_param("world_name".to_string(), serde_json::json!(world_name));
+
+    // Populate mission/step/archetype context from entity state
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::PlayerLoaded,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    if resolved.actions.is_empty() {
+        tracing::info!(entity_id, player_id, %world_name, "fire_player_loaded: no chains matched");
+    } else {
+        for (chain_id, action) in &resolved.actions {
+            tracing::info!(entity_id, player_id, %world_name, chain_id, action = ?action, "fire_player_loaded: matched action");
+        }
+    }
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+
+    // Diagnostic: confirm mission 622 state after chain execution
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        let m622_active = entity.missions.get_mission(622)
+            .map_or(false, |m| m.status == 1);
+        tracing::info!(
+            entity_id, player_id, %world_name,
+            mission_622_active = m622_active,
+            total_missions = entity.missions.count(),
+            "fire_player_loaded: post-execute state"
+        );
+    }
+}
+
+/// Fire the `DialogOpen` event when a dialog is displayed to a player.
+pub async fn fire_dialog_open(
+    entity_id: u32,
+    player_id: i32,
+    dialog_id: i32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("dialog_id".to_string(), serde_json::json!(dialog_id));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+    }
+
+    // Diagnostic: show mission 622 state and chain count before resolution
+    let mission_status = ctx.params.get("mission_622_status")
+        .and_then(|v| v.as_str())
+        .unwrap_or("<not set>");
+    tracing::info!(
+        entity_id, player_id, dialog_id,
+        mission_622_status = %mission_status,
+        dialog_open_chains = engine.chains_for_trigger(&TriggerType::DialogOpen),
+        "fire_dialog_open: resolving"
+    );
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::DialogOpen,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+
+    tracing::info!(
+        entity_id, dialog_id,
+        matched_actions = resolved.actions.len(),
+        "fire_dialog_open: resolved"
+    );
+
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire `OnInteractTag` event when a player interacts with a tagged entity.
+///
+/// Returns `true` if a content chain handled the interaction (caller should
+/// NOT fall through to the default `handle_interact()` logic).
+pub async fn fire_interact_tag(
+    entity_id: u32,
+    player_id: i32,
+    tag: &str,
+    target_entity_id: u32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) -> bool {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("entity_tag".to_string(), serde_json::json!(tag));
+    ctx.set_param("target_entity_id".to_string(), serde_json::json!(target_entity_id));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::InteractTag,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: Some(cimmeria_common::EntityId(target_entity_id as i32)),
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    let matched = !resolved.actions.is_empty();
+    if matched {
+        tracing::info!(entity_id, player_id, %tag, actions = resolved.actions.len(), "fire_interact_tag: matched");
+        executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+    } else {
+        tracing::debug!(entity_id, %tag, "fire_interact_tag: no chains matched");
+    }
+    matched
+}
+
+/// Fire `OnInteractTemplate` event when a player interacts with a templated entity.
+///
+/// Returns `true` if a content chain handled the interaction.
+pub async fn fire_interact_template(
+    entity_id: u32,
+    player_id: i32,
+    template_name: &str,
+    target_entity_id: u32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) -> bool {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("template_name".to_string(), serde_json::json!(template_name));
+    ctx.set_param("target_entity_id".to_string(), serde_json::json!(target_entity_id));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::InteractTemplate,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: Some(cimmeria_common::EntityId(target_entity_id as i32)),
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    let matched = !resolved.actions.is_empty();
+    if matched {
+        tracing::info!(entity_id, player_id, %template_name, actions = resolved.actions.len(), "fire_interact_template: matched");
+        executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+    } else {
+        tracing::debug!(entity_id, %template_name, "fire_interact_template: no chains matched");
+    }
+    matched
+}
+
+/// Fire the `EntityDeath` event when an NPC is killed.
+///
+/// This triggers content chains that track kill counts for mission progression
+/// (e.g., chains 1085-1086: kill Hallway01_Guard → increment counter → complete mission 681).
+pub async fn fire_entity_death(
+    killer_entity_id: u32,
+    player_id: i32,
+    entity_tag: &str,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(killer_entity_id as i32));
+    ctx.set_param("entity_tag".to_string(), serde_json::json!(entity_tag));
+
+    if let Some(entity) = space_mgr.get_entity(killer_entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::EntityDeath,
+        source_entity: Some(cimmeria_common::EntityId(killer_entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    if !resolved.actions.is_empty() {
+        tracing::info!(
+            killer = killer_entity_id, player_id, %entity_tag,
+            actions = resolved.actions.len(), "fire_entity_death: matched"
+        );
+    }
+    executor::execute_actions(resolved, killer_entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire the `RegionEnter` event when the client enters a Kismet region.
+///
+/// `region_tag` is the DB `point_sets.name` value (e.g., "Castle_Cellblock.Region2")
+/// which doubles as the content engine trigger key. The Python reference fires
+/// `client_hinted_region::{tag}` — our content engine matches on `region_key`.
+pub async fn fire_enter_region(
+    entity_id: u32,
+    player_id: i32,
+    region_tag: &str,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("region_key".to_string(), serde_json::json!(region_tag));
+
+    let world_name = space_mgr.get_entity_world_name(entity_id)
+        .unwrap_or_else(|| "Unknown".to_string());
+    ctx.set_param("world_name".to_string(), serde_json::json!(&world_name));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::RegionEnter,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    if !resolved.actions.is_empty() {
+        tracing::info!(entity_id, player_id, %region_tag, actions = resolved.actions.len(), "fire_enter_region: matched");
+    } else {
+        tracing::debug!(entity_id, %region_tag, "fire_enter_region: no chains matched");
+    }
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire the `RegionExit` event when the client exits a Kismet region.
+///
+/// See [`fire_enter_region`] for parameter documentation.
+pub async fn fire_exit_region(
+    entity_id: u32,
+    player_id: i32,
+    region_tag: &str,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("region_key".to_string(), serde_json::json!(region_tag));
+
+    let world_name = space_mgr.get_entity_world_name(entity_id)
+        .unwrap_or_else(|| "Unknown".to_string());
+    ctx.set_param("world_name".to_string(), serde_json::json!(&world_name));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::RegionExit,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    if !resolved.actions.is_empty() {
+        tracing::info!(entity_id, player_id, %region_tag, actions = resolved.actions.len(), "fire_exit_region: matched");
+    } else {
+        tracing::debug!(entity_id, %region_tag, "fire_exit_region: no chains matched");
+    }
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire `OnDialogChoice` event when a player clicks a dialog button.
+pub async fn fire_dialog_choice(
+    entity_id: u32,
+    player_id: i32,
+    dialog_id: i32,
+    button_id: i32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("dialog_id".to_string(), serde_json::json!(dialog_id));
+    ctx.set_param("button_id".to_string(), serde_json::json!(button_id));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::DialogChoice,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    if !resolved.actions.is_empty() {
+        tracing::info!(entity_id, player_id, dialog_id, actions = resolved.actions.len(), "fire_dialog_choice: matched");
+    } else {
+        tracing::debug!(entity_id, dialog_id, "fire_dialog_choice: no chains matched");
+    }
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire `OnItemUse` event when a player uses an inventory item.
+pub async fn fire_item_use(
+    entity_id: u32,
+    player_id: i32,
+    item_id: i32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let mut ctx = ExecutionContext::new()
+        .with_source(cimmeria_common::EntityId(entity_id as i32));
+    ctx.set_param("item_id".to_string(), serde_json::json!(item_id));
+
+    if let Some(entity) = space_mgr.get_entity(entity_id) {
+        populate_mission_context(entity, &mut ctx);
+        if let Some(archetype_id) = entity.archetype_id {
+            ctx.set_param("archetype".to_string(), serde_json::json!(archetype_id));
+        }
+    }
+
+    let event = TriggerEvent {
+        trigger_type: TriggerType::ItemUse,
+        source_entity: Some(cimmeria_common::EntityId(entity_id as i32)),
+        target_entity: None,
+        params: ctx.params.clone(),
+    };
+
+    let resolved = engine.resolve_event(&event, &ctx);
+    let matched = !resolved.actions.is_empty();
+    if matched {
+        tracing::info!(entity_id, player_id, item_id, actions = resolved.actions.len(), "fire_item_use: matched");
+    } else {
+        tracing::debug!(entity_id, item_id, "fire_item_use: no chains matched");
+    }
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
+
+/// Fire a content chain directly by ID, bypassing trigger matching.
+///
+/// Used for minigame victory callbacks — the chain has no trigger row,
+/// it's invoked explicitly by the minigame result handler.
+pub async fn fire_chain_by_id(
+    chain_id: i64,
+    entity_id: u32,
+    player_id: i32,
+    engine: &ChainEngine,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let actions = engine.get_chain_actions(chain_id);
+    if actions.is_empty() {
+        tracing::warn!(chain_id, entity_id, "fire_chain_by_id: chain not found or has no actions");
+        return;
+    }
+
+    tracing::info!(chain_id, entity_id, action_count = actions.len(), "fire_chain_by_id: executing");
+    let resolved = cimmeria_content_engine::chain::ResolvedActions {
+        actions: actions.into_iter().map(|a| (chain_id, a)).collect(),
+    };
+    executor::execute_actions(resolved, entity_id, player_id, tx, space_mgr).await;
+}
